+++ packages/web-api/test/Unit/WebApi/LoginAttemptRepositorySpec.hs
@@ -85,5 +85,6 @@ databaseConfig =
       databasePort = 5432,
       databaseName = "web_api_test",
       databaseUser = "web_api_runtime",
-      databasePassword = "password"
+      databasePassword = "password",
+      databaseConnectTimeoutSeconds = 10
     }