+++ packages/web-api/test/Unit/WebApi/MfaEnrollmentSessionSpec.hs
@@ -163,7 +163,8 @@ databaseConfig =
       databasePort = 5432,
       databaseName = "web_api_test",
       databaseUser = "web_api_runtime",
-      databasePassword = "password"
+      databasePassword = "password",
+      databaseConnectTimeoutSeconds = 10
     }
 
 accountId :: AccountId